AgroHelpDesk/backend/requirements.txt
fastapi
uvicorn[standard]
httpx
openai
python-dotenv
pydantic
pydantic-settings
azure-search-documents
azure-identity
azure-keyvault-secrets
azure-communication-chat
azure-communication-identity
aiohttp
requests
semantic-kernel>=1.0.0
azure-ai-inference